% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs.prototype-package.R
\docType{package}
\name{pbs.prototype-package}
\alias{pbs.prototype}
\alias{pbs.prototype-package}
\title{pbs.prototype: A prototype plant breeding simulation with a large language model using edibble and AlphaSimR}
\description{
This package is only a prototype to demonstrate the possibility of leveraging large language models to automate the specification of simulation using AlphaSimR with a given experimental design via edibble.
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

}
\keyword{internal}
