% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosim.R
\name{simulate_plant_experiment}
\alias{simulate_plant_experiment}
\title{Auto-simulate the records for plant breeding with AlphaSimR}
\usage{
simulate_plant_experiment(
  design = NULL,
  chat = ellmer::chat_openai(),
  context = NULL
)
}
\description{
This function would read in an edibble experiment (that must include record factors)
and infers the context to determine what the appropriate values or distributions
that the record factor takes place using the power of large language models (LLM).
This function is similar in spirit with `edibble::autofill_rcrds()`, however,
does not require the manual specification of providing expected values and
the simulation scheme is inferred from some context provided by LLM.
}
